cuda/requirements.txt
fastapi==0.103.0
numpy==1.24.4
opencv_contrib_python==4.6.0.66
opencv_python_headless==4.6.0.66
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
# torch==2.7.0
# torchvision==0.22.0
cn-clip==1.5.1
RapidOCR==2.0.6
